force each attribute method definition individually to avoid migration hell
if you migrate while having the method enumerated in the model it would
throw and an error from the alias_method_chain. doing it individually avoids this
because it doesn't rely on the method existing after
define_method_attributes call. which it won't for methods being added in
the migration